fix(sshkeys-modal): refactor sshkeys-modal
